% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.ending.point.R
\name{function.ending.point}
\alias{function.ending.point}
\title{function.ending.point}
\usage{
function.ending.point(all.chars, beginning.index, ...)
}
\arguments{
\item{all.chars}{TBD}

\item{beginning.index}{TBD}

\item{...}{additional arguments to be passed.}
}
\description{
Internal function
}
