% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.fn.calls.R
\name{translate.fn.calls}
\alias{translate.fn.calls}
\title{translate.fn.calls}
\usage{
translate.fn.calls(the.statement, function.name = "get(",
  envir = .GlobalEnv, ...)
}
\arguments{
\item{the.statement}{a character value expressing a data.table calculation, such as \code{dt[Age < 50, .(mean(Income)), by = "Region"]}. When the.statement includes variables that are called by reference to a naming constant using get() or eval(), then these statements are translated into the names of the variables for substantial improvements in speed.}

\item{function.name}{translation basis either "get(" or "eval(" to perform the corresponding function. Set as "get(" as default.}

\item{envir}{The environment in which the calculation takes place, with the global environment .GlobalEnv set as the default.}

\item{...}{additional arguments to be passed.}
}
\description{
Internal function
}
